feat: outline
