+++ templates/vue-demo-store/components/account/AccountOrderItem.vue
@@ -12,42 +12,18 @@ const itemQuantity = computed(() => props.orderItem?.quantity);
 const productOptions = computed(() =>
   isProduct(props.orderItem.payload) ? props.orderItem.payload.options : [],
 );
+
+const LineItemType = computed(() => {
+  switch (props.orderItem.type) {
+    case 'product':
+      return defineAsyncComponent(
+          () => import("./order/LineItemProduct.vue"),
+      );
+  }
+
+})
 </script>
 
 <template>
-  <div
-    v-if="!isPromotion"
-    class="h-24 w-24 flex-shrink-0 overflow-hidden rounded-md border border-secondary-200"
-  >
-    <img
-      :src="getMainImageUrl(orderItem)"
-      alt="Salmon orange fabric pouch with match zipper, gray zipper pull, and adjustable hip belt."
-      class="h-full w-full object-cover object-center"
-    />
-  </div>
-
-  <div class="ml-4 flex flex-1 flex-col">
-    <div>
-      <div
-        class="flex justify-between text-base font-medium text-secondary-900"
-      >
-        <h3>
-          {{ orderItem.label }}
-        </h3>
-        <p class="ml-4">{{ itemRegularPrice }} EUR</p>
-      </div>
-      <p class="mt-1 text-sm text-secondary-500">
-        <span
-          v-for="option in productOptions"
-          :key="option.group.name"
-          class="mr-2"
-        >
-          {{ option.group.name }}: {{ option.option }}
-        </span>
-      </p>
-    </div>
-    <div class="flex flex-1 items-end justify-between text-sm">
-      <p class="text-secondary-500">Qty {{ itemQuantity }}</p>
-    </div>
-  </div>
+  <LineItemType :line-item="orderItem" />
 </template>
