+++ templates/vue-demo-store/i18n/de-DE/cart.json
@@ -11,6 +11,7 @@
     "emptyCartLabel": "Ihr Warenkorb ist leer!",
     "messages": {
       "addedToCart": "{p} wurde dem Warenkorb hinzugefügt."
-    }
+    },
+    "promotion": "Promotion"
   }
 }
