+++ templates/vue-demo-store/i18n/en-GB/cart.json
@@ -11,6 +11,7 @@
     "emptyCartLabel": "Your cart is empty!",
     "messages": {
       "addedToCart": "{p} has been added to cart."
-    }
+    },
+    "promotion": "promotion"
   }
 }
