+++ templates/vue-demo-store/i18n/pl-PL/cart.json
@@ -11,6 +11,7 @@
     "emptyCartLabel": "Twój koszyk jest pusty!",
     "messages": {
       "addedToCart": "{p} został dodany do koszyka."
-    }
+    },
+    "promotion": "promocja"
   }
 }
